Phase 3: consolidate serializers into ctrack/api/serializers/
Move all inline serializers out of the API view modules into one
module per domain under ctrack/api/serializers/. Fold the stray
top-level data_serializer.py and series_serializer.py into a new
serializers/common.py (SeriesSerializer, LoadDataSerializer).

Add file-upload validation (size ceiling + extension allow-list) to
LoadDataSerializer, which backs AccountViewSet.load and
RecurringPaymentViewSet.loadpdf, with tests.

No model/migration changes; OpenAPI schema diff is empty.
