+++ ctrack/api/budget_entry.py
@@ -2,16 +2,12 @@
 """
 import logging
 
-from rest_framework import serializers, viewsets
+from rest_framework import viewsets
+from ctrack.api.serializers.budget_entry import BudgetEntrySerializer
 from ctrack.models import BudgetEntry
 
 
 logger = logging.getLogger(__name__)
-
-class BudgetEntrySerializer(serializers.HyperlinkedModelSerializer):
-    class Meta:
-        model = BudgetEntry
-        fields = ('url', 'id', 'pretty_name', 'amount', 'valid_from', 'valid_to', 'categories')
 
 class BudgetEntryViewSet(viewsets.ModelViewSet):
     queryset = BudgetEntry.objects.all()
